modified CurrentPromptModal to handle failing test cases
